CLN: added io.api for i/o importing functions
     moved excel functionaility out of io.parsers to io.excel
     added read_excel top-level function
       aliases from pandas.io.excel
     added read_stata top-level function, to_stata DataFrame method
       aliases from pandas.io.stata
       removed read_dta (replace by read_stata)
     added read_sql   top-level function, to_sql DataFrame method
       aliases from pandas.io.sql

DOC: doc updates for all the above and intro section to io.rst
