+++ doc/source/v0.10.0.txt
@@ -1,5 +1,10 @@
 .. _whatsnew_0100:
 
+.. ipython:: python
+   :suppress:
+
+   from StringIO import StringIO
+
 v0.10.0 (December 17, 2012)
 ---------------------------
