+++ pandas/__init__.py
@@ -28,12 +28,8 @@ from pandas.core.api import *
 from pandas.sparse.api import *
 from pandas.stats.api import *
 from pandas.tseries.api import *
+from pandas.io.api import *
 
-from pandas.io.parsers import (read_csv, read_table, read_clipboard,
-                               read_fwf, to_clipboard, ExcelFile,
-                               ExcelWriter)
-from pandas.io.pytables import HDFStore, Term, get_store, read_hdf
-from pandas.io.html import read_html
 from pandas.util.testing import debug
 
 from pandas.tools.describe import value_range
