+++ pandas/core/generic.py
@@ -491,6 +491,10 @@ class PandasObject(object):
         from pandas.io import pytables
         return pytables.to_hdf(path_or_buf, key, self, **kwargs)
 
+    def to_clipboard(self):
+        from pandas.io import parsers
+        parsers.to_clipboard(self)
+
 # install the indexerse
 for _name, _indexer in indexing.get_indexers_list():
     PandasObject._create_indexer(_name,_indexer)
